# adds support for immutable datastructures