--- a/package.json
+++ b/package.json
@@ -23,6 +23,7 @@
     "gulp-plumber": "^1.0.1",
     "gulp-rename": "^1.2.2",
     "gulp-uglify": "^1.2.0",
+    "immutable": "^3.7.6",
     "nodangel": "^1.3.8",
     "yargs": "^3.15.0"
   },