How do you feel about adding support for [`Immutable.js`][i]?
I created this POC that it would work, if you like it I can add more tests and documentation.
It would be great to see this feature in sift.
We currently do `Immutable.fromJS(list.toJS().filter(sift(query)));` which is extremely slow (because of the convertion from and to js).

The only thing I changed is how values are accessed on objects/arrays.

[i]: http://facebook.github.io/immutable-js/