@stoeffel 2016-01-14
I tried it in our app and it wasn't really faster. So feel free to close this PR.