@stoeffel 2016-01-18
> I tried it in our app and it wasn't really faster.

I think I was somehow confused last week :blush: 

I ran a benchmark and this are the results:
```
{"num":{"$lt":5}} x 10 items (converting from/toJS) x 12,571 ops/sec ±2.09% (83 runs sampled)
{"num":{"$lt":5}} x 10 items (no convertion) x 90,175 ops/sec ±1.82% (81 runs sampled)
{"num":{"$lt":5}} x 40 items (converting from/toJS) x 2,781 ops/sec ±1.87% (84 runs sampled)
{"num":{"$lt":5}} x 40 items (no convertion) x 36,770 ops/sec ±1.31% (94 runs sampled)
{"num":{"$lt":5}} x 120 items (converting from/toJS) x 953 ops/sec ±2.70% (79 runs sampled)
{"num":{"$lt":5}} x 120 items (no convertion) x 13,871 ops/sec ±4.29% (81 runs sampled)
{"num":{"$lt":5}} x 320 items (converting from/toJS) x 387 ops/sec ±4.64% (77 runs sampled)
{"num":{"$lt":5}} x 320 items (no convertion) x 6,257 ops/sec ±2.27% (91 runs sampled)
Fastest is {"num":{"$lt":5}} x 10 items (no convertion)
```
```js
// all tests with (converting from/toJS) are executing:
Immutable.fromJS(immuList.toJS().filter(sift(query)));

// all tests with (no convertion) are executing:
array.filter(sift(query));
```

So in conclusion this PR does give a big performance improvement.
Let me know what you think about this changes. Do you want me to change anything? I can include my benchmark in this PR if you like.

Thanks.