@stoeffel 2016-01-18
[Link to the benchmark source](https://github.com/stoeffel/sift.js/blob/immutable-benchmark/benchmark/immutable.js#L113)