@adrai 2016-01-19
Yes!!!! This would be a super performance boost! Would be great if this could be merged soon.