@stoeffel 2016-01-19
The performance difference between the current master and this PR is marginal (for none-immutable objects).

Master:
```
$ node benchmark/index.js
{"num":{"$lt":5}} x 10 items x 170,966 ops/sec ±0.78% (95 runs sampled)
{"num":{"$lt":5}} x 40 items x 62,535 ops/sec ±1.62% (96 runs sampled)
{"num":{"$lt":5}} x 120 items x 23,511 ops/sec ±0.59% (91 runs sampled)
{"num":{"$lt":5}} x 320 items x 8,796 ops/sec ±1.05% (92 runs sampled)
{"num":{"$lt":5}} x 800 items x 3,615 ops/sec ±1.39% (92 runs sampled)
{"num":{"$lt":5}} x 1920 items x 1,550 ops/sec ±1.85% (92 runs sampled)
{"num":{"$lt":5}} x 4480 items x 660 ops/sec ±1.63% (89 runs sampled)
{"num":{"$lt":5}} x 10240 items x 306 ops/sec ±0.62% (91 runs sampled)
{"num":{"$lt":5}} x 23040 items x 140 ops/sec ±0.94% (78 runs sampled)
Fastest is {"num":{"$lt":5}} x 10 items
```

With support for immutables
```
$ node benchmark/index.js
{"num":{"$lt":5}} x 10 items x 159,074 ops/sec ±1.68% (91 runs sampled)
{"num":{"$lt":5}} x 40 items x 60,476 ops/sec ±1.51% (92 runs sampled)
{"num":{"$lt":5}} x 120 items x 22,724 ops/sec ±1.22% (95 runs sampled)
{"num":{"$lt":5}} x 320 items x 8,207 ops/sec ±2.31% (91 runs sampled)
{"num":{"$lt":5}} x 800 items x 3,561 ops/sec ±0.74% (97 runs sampled)
{"num":{"$lt":5}} x 1920 items x 1,492 ops/sec ±1.55% (93 runs sampled)
{"num":{"$lt":5}} x 4480 items x 671 ops/sec ±0.58% (95 runs sampled)
{"num":{"$lt":5}} x 10240 items x 284 ops/sec ±2.82% (85 runs sampled)
{"num":{"$lt":5}} x 23040 items x 129 ops/sec ±4.51% (79 runs sampled)
Fastest is {"num":{"$lt":5}} x 10 items
```