@stoeffel 2016-01-20
@crcn would be great to get a feedback on this.