@crcn 2016-01-20
Looks like a useful addition that fits other cases. I may modify this a bit in the future to provide support for other array-like data structures. Thanks!