@crcn 2016-01-20
Updated on NPM. New version is `3.2.0` for this feature.